Add useless cast warning & fix errors

+++ libr/include/r/clock.h
@@ -31,8 +31,8 @@
 extern "C" {
 #endif
 
-#define MICROSECONDS_IN_SECOND  (double)1000000.0
-#define NANOSECONDS_IN_SECOND  (double)1000000000.0
+#define MICROSECONDS_IN_SECOND  1000000.0
+#define NANOSECONDS_IN_SECOND  1000000000.0
 
 // Micro-seconds since UNIX epoch
 uint64_t clock_timestamp();
